+++ CHANGELOG.md
@@ -21,6 +21,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 - `setFlags`: function to replace `flagManual`
 - `flagUniLOF`: added defaultly applied correction to mitigate phenomenon of overflagging at relatively steep data value slopes. (parameter `slope_correct`).
 - `History`: added option to change aggregation behavior
+- Translation scheme `FloatScheme`
 ### Changed
 - `flagPattern` uses *fastdtw* package now to compute timeseries distances
 - `SaQC.flags` always returns a `DictOfSeries`